+++ Demo/macOS/EditorScreen.swift
@@ -45,7 +45,11 @@ struct EditorScreen: View {
 private extension EditorScreen {
 
     var editor: some View {
-        RichTextEditor(text: $text, context: context) {
+        RichTextEditor(
+            text: $text,
+            context: context,
+            config: .init(isScrollingEnabled: false)
+        ) {
             $0.textContentInset = CGSize(width: 10, height: 20)
         }
         .frame(minWidth: 400)